requirements.txt
streamlit
pypdf
python-dotenv

# LangChain + OpenRouter
langchain
langchain-openai

# Search
tavily-python
